requirements.txt
langchain==0.1.7
langchain-community==0.0.19
langchain-google-genai==0.0.11
chromadb==0.4.24
google-generativeai==0.3.0
python-dotenv==1.0.0
colorama==0.4.6
